requirements.txt
python-telegram-bot[job-queue]==21.7
python-dotenv==1.0.0
schedule==1.2.1
cachetools==5.3.3
pymongo==4.6.1
requests==2.31.0
aiohttp==3.9.3
colorama==0.4.6
pyfiglet==0.8.post1
tabulate==0.9.0
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0